requirements.txt
numpy 
pandas 
python-dotenv   #helps in reading the .env file environment var
mysql-connector-python
pymysql  #pymylsql is package contains pure-python MySQL client linrary
scikit-learn
#-e .